+++ readium/lcp/src/main/java/org/readium/r2/lcp/license/model/components/lsd/Event.kt
@@ -1,27 +1,37 @@
 /*
- * Module: r2-lcp-kotlin
- * Developers: Aferdita Muriqi
- *
- * Copyright (c) 2019. Readium Foundation. All rights reserved.
- * Use of this source code is governed by a BSD-style license which is detailed in the
- * LICENSE file present in the project repository where this source code is maintained.
+ * Copyright 2026 Readium Foundation. All rights reserved.
+ * Use of this source code is governed by a BSD-style license
+ * available in the top-level LICENSE file of the project.
  */
 
-@file:OptIn(InternalReadiumApi::class)
+@file:OptIn(InternalReadiumApi::class, ExperimentalReadiumApi::class)
 
 package org.readium.r2.lcp.license.model.components.lsd
 
 import kotlin.time.Instant
+import kotlinx.serialization.SerialName
+import kotlinx.serialization.Serializable
+import kotlinx.serialization.encodeToString
 import org.json.JSONObject
+import org.readium.r2.lcp.license.model.LcpJson
+import org.readium.r2.shared.ExperimentalReadiumApi
 import org.readium.r2.shared.InternalReadiumApi
-import org.readium.r2.shared.extensions.optNullableString
-import org.readium.r2.shared.extensions.toInstant
+import org.readium.r2.shared.util.KotlinInstantSerializer
 
-public data class Event(val json: JSONObject) {
-    val type: String = json.optNullableString("type") ?: ""
-    val name: String = json.optNullableString("name") ?: ""
-    val id: String = json.optNullableString("id") ?: ""
-    val date: Instant? = json.optNullableString("timestamp")?.toInstant()
+@Serializable
+public data class Event(
+    @SerialName("type")
+    val type: String = "",
+    @SerialName("name")
+    val name: String = "",
+    @SerialName("id")
+    val id: String = "",
+    @SerialName("timestamp")
+    @Serializable(with = KotlinInstantSerializer::class)
+    val date: Instant? = null,
+) {
+    @Deprecated("Use kotlinx.serialization to serialize the object")
+    val json: JSONObject get() = JSONObject(LcpJson.encodeToString(this))
 
     public enum class EventType(public val value: String) {
         Register("register"),