+++ readium/shared/src/main/java/org/readium/r2/shared/util/KotlinInstantSerializer.kt
@@ -1,0 +1,35 @@
+/*
+ * Copyright 2026 Readium Foundation. All rights reserved.
+ * Use of this source code is governed by a BSD-style license
+ * available in the top-level LICENSE file of the project.
+ */
+
+@file:OptIn(InternalReadiumApi::class)
+
+package org.readium.r2.shared.util
+
+import kotlin.time.Instant
+import kotlinx.serialization.KSerializer
+import kotlinx.serialization.descriptors.PrimitiveKind
+import kotlinx.serialization.descriptors.PrimitiveSerialDescriptor
+import kotlinx.serialization.descriptors.SerialDescriptor
+import kotlinx.serialization.encoding.Decoder
+import kotlinx.serialization.encoding.Encoder
+import org.readium.r2.shared.ExperimentalReadiumApi
+import org.readium.r2.shared.InternalReadiumApi
+import org.readium.r2.shared.extensions.toInstant
+
+@ExperimentalReadiumApi
+public object KotlinInstantSerializer : KSerializer<Instant> {
+    override val descriptor: SerialDescriptor =
+        PrimitiveSerialDescriptor("kotlin.time.Instant", PrimitiveKind.STRING)
+
+    override fun deserialize(decoder: Decoder): Instant {
+        val string = decoder.decodeString()
+        return string.toInstant() ?: throw IllegalArgumentException("Invalid Instant: $string")
+    }
+
+    override fun serialize(encoder: Encoder, value: Instant) {
+        encoder.encodeString(value.toString())
+    }
+}